gpg: New list-options show-pref and show-pref-verbose.
* g10/options.h (LIST_SHOW_PREF): New.
(LIST_SHOW_PREF_VERBOSE): New.
* g10/gpg.c (parse_list_options): Add new options.
* g10/keyedit.c (show_prefs): Factor code out to ...
* g10/keylist.c (show_preferences): new.
(list_keyblock_print): Call show_preferences.
--
Backported-from-master: 811cfa34cb3e7166f0cf1f94565504dee21cd9f5

+++ g10/cipher.c
@@ -436,9 +436,9 @@ do_ocb_flush (cipher_filter_context_t *cfx, iobuf_t a, byte *buf, size_t size)
           size_t n1 = cfx->chunksize - (cfx->chunklen + cfx->buflen);
           finalize = 1;
           if (DBG_FILTER)
-            log_debug ("chunksize %"PRIu64" reached;"
+            log_debug ("chunksize %zu reached;"
                        " cur buflen=%zu using %zu of %zu\n",
-                       cfx->chunksize, cfx->buflen,
+                       (size_t)cfx->chunksize, cfx->buflen,
                        n1, n);
           n = n1;
         }
